refactor: SuccessConfirmPage 에서 useLocation 으로 상태를 전달받아 결제 금액 띄워지도록 구조 변경
